+++ libs/sdk-core/src/lib/db/db.models.ts
@@ -219,6 +219,16 @@ export interface LearningConfiguration {
   update: boolean;
 }
 
+export function normalizeSchemaProperty(config: LearningConfigurations): LearningConfigurations {
+  Object.values(config).forEach((item) => {
+    if (item.schemas?.['title'] && item.schemas?.['type']) {
+      item.schema = item.schemas as unknown as LearningConfigurationSchema;
+      item.schemas = undefined;
+    }
+  });
+  return config;
+}
+
 export const USER_PROMPTS = 'user_prompts';
 export const SUMMARY_PROMPT = 'summary_prompt';